requirements.txt
pytesseract
pdf2image
Pillow
transformers
torch
streamlit
langdetect
sacremoses
sentencepiece
PyMuPDF
